+++ changelog.md
@@ -15,6 +15,7 @@
       - Add standard object support to add money, gems, and standard objects
       - Add argument support to objects and monsters to allow passing arguments when object is created
       - Hide water properties unless waterdrink flag enabled
+      - Add monster type basic vendor and barkeep property support
     - Virtual area editor: Add ctrl+click selection system
 - **Fixed:**
   - Increase progress dialog height